engine/v1_backup/requirements.txt
tastytrade>=12.0.0
python-kraken-sdk>=3.0.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
